Changed ShenandoahLogFileOutput to independent class and added copyrights
